move components to separate folder, update routing
